fix(web):  add a loaders folder for each page
