+++ src/loaders/feedback/index.ts
@@ -1,0 +1,25 @@
+import { fetcher } from "../../utils";
+
+export const progressBarLoader = async () => {
+  const vaccinationData = await fetcher("vaccination/vax_state.csv");
+  const populationData = await fetcher("static/population.csv");
+
+  const filteredVaccinationData = vaccinationData
+    .slice(-17)
+    .filter((row: any) => !!row.state);
+  const filteredPopulationdata = populationData.filter(
+    (row: any) => !!row.state
+  );
+
+  const data = filteredVaccinationData.map((row: any) => {
+    const state: any = filteredPopulationdata.find(
+      ({ state }: any) => state === row.state
+    );
+    return {
+      group: row["state"],
+      value: (+row["cumul_full"] / +state["pop"]) * 100,
+    };
+  });
+
+  return data;
+};
